remove old version of logrus (github.com/Sirupsen/logrus)
add new version of logrus (github.com/sirupsen/logrus)
lock logrus version to 1.0.6
add logger/logger.go for custom TextFormatter used with logrus

first part of solving #624