@Dieterbe 2018-09-20
some notes on the output formats:
metrictank:
```
2018-09-20 09:46:11.769 [WARNING] [metrictank] stats dialing localhost:2003 failed: dial tcp 127.0.0.1:2003: connect: connection refused. will retry
2018-09-20 09:46:14.386 [WARNING] [metrictank] kafka-cluster: Processing metricPersist backlog has taken too long, giving up lock after 1m0s.
2018-09-20 09:46:14.386 [INFO] [metrictank] carbon-in: listening on :2003/tcp
[Sarama] 2018/09/20 09:46:14 Connected to broker at kafka:9092 (registered as #0)
2018-09-20 09:46:14.394 [INFO] [metrictank] kafkamdm: consuming from mdm:3 from offset 0
2018-09-20 09:46:14.395 [INFO] [metrictank] kafkamdm: consuming from mdm:4 from offset 0
```

tools:
```
./build/mt-index-cat -addr http://localhost:6063 -from 60min cass -hosts localhost:9042 -schema-file scripts/config/schema-idx-cassandra.toml dump | less
2018-09-20 12:45:51.667 [INFO] [mt-index-cat] cassandra-idx: ensuring that keyspace metrictank exist.
2018-09-20 12:45:51.668 [INFO] [mt-index-cat] cassandra-idx: ensuring that table metric_idx exist.
```

1) it's unfortunate that libraries such as sarama still do their own logging. but that's outside scope of this PR
2) i don't think we need to print which binary is running. as that is obvious from context (e.g. pod running in, command that was run, etc). in fact for example in our logging platform we'll know which program is running as it'll be a label that we can query on. it doesn't need to be part of the output lines.