@Dieterbe 2018-09-20
can you compare benches before and after this change?
you can use benchstat to compare. make sure to run all benches several time, both pre and post your changeset, to get accurate results. if the results are still noisy, you may have to disable cpu frequency scaling etc on your machine