@robert-milan 2018-09-23
I've been running benchmarks on master and this branch. I have disabled cpu frequency scaling and set my governor to performance, but I am still getting consistently inconsistent results specifically with the BenchmarkAggregate10k_* and BenchmarkGroupByTags10k_* series of benchmarks. Most of the other benchmarks are within tolerance of each other, but I am not very confident using the current numbers I have as a base for comparison. I am also receiving the same benchmarking failures on both branches. I have attached a zip of the results here.
[benchmarks.zip](https://github.com/grafana/metrictank/files/2409183/benchmarks.zip)