@Dieterbe 2018-10-02
we haven't run the benchmarks in a while apparently, i didn't realize some of them are broken.
addressing via #1077 

to compare performance, will be most interesting to look at some methods that do a lot of Debug calls, especially those that have their LogLevel shields removed. e.g.  
AggMetric.SyncChunkSaveState
AggMetric.Get
AggMetric.addAggregators
AggMetric.persist
AggMetric.add

`add()` is a good one, and is well represented in these benchmarks:

BenchmarkAggMetrics1000Metrics1Day
BenchmarkAggMetrics1kSeries2Chunks1kQueueSize
BenchmarkAggMetrics10kSeries2Chunks10kQueueSize
BenchmarkAggMetrics100kSeries2Chunks100kQueueSize