@Dieterbe 2018-10-02
with #1077 integrated in both master and this branch, I get:
```
benchstat before-new.txt after-new.txt
name          old time/op  new time/op  delta
AggMetricAdd   503ns ± 1%   520ns ± 2%  +3.44%  (p=0.000 n=10+10)
```
so a slight slowdown. no biggie